requirements.txt
keyboard
pyautogui
pywin32
pystray
Pillow
openai
google-generativeai>=0.3.0
PyInstaller 
wheel
pyinstaller-hooks-contrib
pycryptodome
pyside6